Add typeshare for automatic server-side type conversion (#664)
