+++ client/src/lib/types.ts
@@ -1,0 +1,23 @@
+/*
+ Generated by typeshare 1.11.0
+*/
+
+export interface Instructor {
+  name: string;
+  nameNgrams?: string;
+  term: string;
+}
+
+export interface Subscription {
+  courseId: string;
+  userId: string;
+}
+
+export interface User {
+  id: string;
+  mail: string;
+}
+
+export interface UserResponse {
+  user?: User;
+}
