+++ client/src/lib/utils.ts
@@ -1,14 +1,16 @@
 import { groupBy } from 'lodash';
 
+import { Instructor } from '../lib/types';
 import { Course } from '../model/Course';
-import { Instructor } from '../model/Instructor';
 import type { Schedule } from '../model/Schedule';
 
 const TERM_ORDER = ['Winter', 'Summer', 'Fall'];
 
 /**
  * Groups course instructors by their terms, but only for current terms.
+ *
  * Creates empty arrays for current terms that have no instructors.
+ *
  * @param {Course} course - The course object containing instructors and terms
  * @returns {Record<string, Instructor[]>} Object mapping terms to arrays of instructors
  */
@@ -33,9 +35,11 @@ export const groupCurrentCourseTermInstructors = (
 
 /**
  * Determines the current and next two academic terms based on the current date.
+ *
  * - May-July: Returns [Summer current, Fall current, Winter next]
  * - August-December: Returns [Fall current, Winter next, Summer next]
  * - January-April: Returns [Fall previous, Winter current, Summer current]
+ *
  * @returns {[string, string, string]} Array of three consecutive terms
  */
 export const getCurrentTerms = (): [string, string, string] => {
@@ -57,9 +61,11 @@ export const getCurrentTerms = (): [string, string, string] => {
 
 /**
  * Determines the current academic term based on the current date.
+ *
  * - May-July: Summer <year>
  * - August-December: Fall <year>
  * - January-April: Winter <year>
+ *
  * @returns string The current term
  */
 export const getCurrentTerm = (): string => {
@@ -80,7 +86,9 @@ export const getCurrentTerm = (): string => {
 
 /**
  * Compares two academic terms for sorting.
+ *
  * Terms are compared first by year, then by season according to TERM_ORDER.
+ *
  * @param {string} a - First term string (e.g., "Fall 2023")
  * @param {string} b - Second term string (e.g., "Winter 2024")
  * @returns {number} Negative if a comes before b, positive if b comes before a, 0 if equal
@@ -93,7 +101,9 @@ export const compareTerms = (a: string, b: string): number => {
 
 /**
  * Sorts an array of academic terms chronologically.
+ *
  * Uses compareTerms to determine order.
+ *
  * @param {string[]} terms - Array of term strings to sort
  * @returns {string[]} Sorted array of terms
  */
@@ -103,8 +113,11 @@ export const sortTerms = (terms: string[]): string[] => {
 
 /**
  * Sorts course schedules by block type and number.
+ *
  * Order priority: Lec > Lab > Seminar > Tut > Conf
+ *
  * Within each type, sorts numerically by block number.
+ *
  * @param {Schedule[]} schedules - Array of course schedules to sort
  * @returns {Schedule[]} Sorted array of schedules
  */
@@ -125,7 +138,9 @@ export const sortSchedulesByBlocks = (schedules: Schedule[]): Schedule[] => {
 
 /**
  * Converts a course ID to a URL-friendly parameter format.
+ *
  * Example: "COMP202" -> "comp-202"
+ *
  * @param {string} courseId - The course ID to convert
  * @returns {string} URL-friendly course ID
  */
@@ -134,6 +149,7 @@ export const courseIdToUrlParam = (courseId: string): string =>
 
 /**
  * Capitalizes the first character of a string.
+ *
  * @param {string} s - The string to capitalize
  * @returns {string} Capitalized string
  */
@@ -142,7 +158,9 @@ export const capitalize = (s: string): string =>
 
 /**
  * Ensures a string ends with a period.
+ *
  * Adds a period if one is not already present.
+ *
  * @param {string} s - The string to punctuate
  * @returns {string} String ending with a period
  */
@@ -153,8 +171,11 @@ const COURSE_CODE_REGEX = /^(([A-Z0-9]){4} [0-9]{3}(D1|D2|N1|N2|J1|J2|J3)?)$/;
 
 /**
  * Validates if a string matches the course code format.
+ *
  * Valid format: 4 alphanumeric chars + space + 3 digits + optional suffix
+ *
  * Suffixes: D1, D2, N1, N2, J1, J2, J3
+ *
  * @param {string} s - The string to validate
  * @returns {boolean} True if string is a valid course code
  */
@@ -163,7 +184,9 @@ export const isValidCourseCode = (s: string): boolean =>
 
 /**
  * Inserts a delimiter between the subject and number portions of a course code.
+ *
  * Example: spliceCourseCode("COMP202", "-") -> "COMP-202"
+ *
  * @param {string} courseCode - The course code to splice
  * @param {string} delimiter - The delimiter to insert
  * @returns {string} Course code with delimiter inserted
@@ -175,6 +198,7 @@ export const spliceCourseCode = (
 
 /**
  * Rounds a number to 2 decimal places.
+ *
  * @param {number} n - Number to round
  * @returns {number} Rounded number
  */
@@ -182,7 +206,9 @@ export const round2Decimals = (n: number): number => Math.round(n * 100) / 100;
 
 /**
  * Performs a true modulo operation (different from JavaScript's % operator).
+ *
  * Always returns a positive number, even when inputs are negative.
+ *
  * @param {number} n - Dividend
  * @param {number} m - Divisor
  * @returns {number} Positive modulo result
@@ -191,6 +217,7 @@ export const mod = (n: number, m: number): number => ((n % m) + m) % m;
 
 /**
  * Custom error class for date-related errors.
+ *
  * @extends Error
  */
 export class InvalidDateError extends Error {
@@ -202,8 +229,11 @@ export class InvalidDateError extends Error {
 
 /**
  * Converts a date to a human-readable "time ago" string.
+ *
  * Handles various time units from seconds to years.
+ *
  * Throws InvalidDateError for null, undefined, invalid formats, or future dates.
+ *
  * @param {Date | string | number | null | undefined} date - Date to convert
  * @returns {string} Human-readable time difference (e.g., "2 hours ago")
  * @throws {InvalidDateError} If date is invalid, missing, or in the future
