--- client/src/model/Instructor.ts
@@ -1,4 +1,0 @@
-export type Instructor = {
-  name: string;
-  term: string;
-};
