--- client/src/model/Subscription.ts
@@ -1,4 +1,0 @@
-export type Subscription = {
-  courseId: string;
-  userId: string;
-};
