--- client/src/model/User.ts
@@ -1,7 +1,0 @@
-export interface User {
-  id: string;
-}
-
-export type UserResponse = {
-  user: User;
-};
