+++ client/src/pages/Instructor.tsx
@@ -11,8 +11,8 @@ import { Layout } from '../components/Layout';
 import { ReviewEmptyPrompt } from '../components/ReviewEmptyPrompt';
 import { useAuth } from '../hooks/useAuth';
 import { repo } from '../lib/repo';
+import type { Instructor as InstructorType } from '../lib/types';
 import { courseIdToUrlParam } from '../lib/utils';
-import type { Instructor as InstructorType } from '../model/Instructor';
 import type { Review } from '../model/Review';
 import { Loading } from './Loading';
 import { NotFound } from './NotFound';
