+++ crates/db/Cargo.toml
@@ -5,9 +5,9 @@ edition = "2021"
 publish = false
 
 [dependencies]
-anyhow = "1.0.98"
+anyhow = { workspace = true }
 bson = { version = "2.14.0", features = ["chrono-0_4"] }
-chrono = "0.4.41"
+chrono = { workspace = true }
 combine = { path = "../combine" }
 futures = "0.3.31"
 itertools = "0.12.1"
@@ -15,13 +15,12 @@ lazy_static = "1.5.0"
 log = "0.4.27"
 model = { path = "../model" }
 mongodb = "2.8.2"
-serde = "1.0.219"
-serde_json = "1.0.140"
+serde = { workspace = true }
+serde_json = { workspace = true }
 stop-words = "0.7.2"
-tokio = { version = "1.45.0", features = ["rt-multi-thread", "macros"] }
+tokio = { workspace = true }
 
 [dev-dependencies]
-chrono = "0.4.41"
 include_dir = "0.7.4"
 pretty_assertions = "1.4.1"
 regex = "1.11.1"
