+++ crates/model/Cargo.toml
@@ -8,5 +8,6 @@ publish = false
 bson = { version = "2.14.0", features = ["chrono-0_4"] }
 combine = { path = "../combine" }
 derivative = "2.2.0"
-serde = { version = "1.0.219", features = ["derive"] }
-serde_json = "1.0.140"
+serde = { workspace = true }
+serde_json = { workspace = true }
+typeshare = { workspace = true }
