+++ crates/model/src/subscription.rs
@@ -2,6 +2,7 @@ use super::*;
 
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
+#[typeshare]
 pub struct Subscription {
   pub course_id: String,
   pub user_id: String,
