+++ justfile
@@ -25,7 +25,7 @@ clippy:
 coverage:
   ./bin/coverage
 
-dev: services
+dev: services typeshare
   concurrently \
     --kill-others \
     --names 'SERVER,CLIENT' \
@@ -42,6 +42,7 @@ dev: services
 
 dev-deps:
   cargo install present
+  cargo install typeshare-cli
   curl -LsSf https://astral.sh/uv/install.sh | sh
 
 e2e:
@@ -107,6 +108,9 @@ services:
 
 test *filter:
   cargo test --all {{filter}}
+
+typeshare:
+  typeshare -l typescript -o client/src/lib/types.ts .
 
 watch +COMMAND='test':
   cargo watch --clear --exec "{{COMMAND}}"
